Deleted an unused Data initialiser
It was causing a warning, that the UnsafeBufferPointer returned a dangling pointer, which means a pointer to a section of memory that could contain anything, not what we would expect, which leads to memory exceptions thrown and crashing the app. We don't want that, but anyway, this was not being used, so it can die.