Removed force-unwraps from RootViewController
There's a crash in thingyPeripheral(_, didChangeStateTo:), and the Organizer does not pin-point where, other than it's in a closure. I think the obvious thing here is to not force-unwrap things since they are obvious sources of crashes. It's not the best solution, but since I can't replicate it and this is not an app that uses a modern codebase or maintained dependencies, taking it step by step seems like the reasonable thing to do to me.